pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "unicorn-viz"
version = "0.1.0"
description = "Modern demoscene visualizer — Razor 1911 vibes, Wayland-native"
requires-python = ">=3.11"
dependencies = [
    "moderngl>=5.10",
    "pysdl2>=0.9.16",
    "pysdl2-dll>=2.28",
    "numpy>=1.26",
    "scipy>=1.12",
    "sounddevice>=0.4.6",
    "python-rtmidi>=1.5",
    "Pillow>=10.0",
]

[project.scripts]
unicorn-viz = "unicornviz.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["unicornviz*"]
